fixed binding tree visualizer route and cleaned spec compiler
