+++ src/binding/utils.ts
@@ -1,0 +1,161 @@
+import { TopLevelSpec } from "vega-lite/build/src/spec";
+
+import { Field } from "vega-lite/build/src/channeldef";
+import { UnitSpec } from "vega-lite/build/src/spec";
+import { Parameter } from "vega-typings/types/runtime/runtime";
+
+// Extract all components from config and their bindings
+export const extractComponentBindings = (config: any): any[] => {
+    // If no config or no bind property, return empty array
+    if (!config || !config.bind || typeof config.bind === 'function') {
+        return [];
+    }
+    
+    // If bind is an array, process each item
+    if (Array.isArray(config.bind)) {
+        return config.bind.flatMap(item => {
+            // If the item is a component with its own bindings, extract those too
+            if (item && typeof item === 'object' && item.bind) {
+                return [item, ...extractComponentBindings(item)];
+            }
+            return item;
+        });
+    }
+    
+    // If bind is a single object with its own bindings
+    if (typeof config.bind === 'object' && config.bind.bind) {
+        return [config.bind, ...extractComponentBindings(config.bind)];
+    }
+    
+    // If bind is a single object without further bindings
+    return [config.bind];
+};
+
+
+export function mergeSpecs(specs: Partial<UnitSpec<Field>>[], rootComponentId: string): TopLevelSpec {
+    // Helper to check if spec has layer/mark
+    const hasLayerOrMark = (spec: any) => {
+        return spec.layer || spec.mark;
+    };
+
+
+    // First merge specs, handling layers
+    const mergedSpec = specs.reduce((merged: any, spec) => {
+        // If either has layer/mark, create layer spec
+        if (hasLayerOrMark(merged) && hasLayerOrMark(spec)) {
+            return {
+                layer: [
+                    merged,
+                    spec
+                ]
+            };
+        }
+
+        // Otherwise merge normally
+        return {
+            ...merged,
+            ...spec,
+            // Concatenate params if they exist
+            params: [
+                ...(merged.params || []),
+                ...(spec.params || [])
+            ]
+        };
+    }, {});
+
+    // Move all params to top level
+    const params: (Parameter)[] = [];
+    const selectParams: (Parameter)[] = [];
+
+    // Function to find base chart layer by rootComponentId
+    const findBaseChartLayer = (obj: any): any | null => {
+        if (!obj || typeof obj !== 'object') return null;
+
+        if (obj.name === rootComponentId) return obj;
+
+        for (const value of Object.values(obj)) {
+            const result = findBaseChartLayer(value);
+            if (result) return result;
+        }
+
+        return null;
+    };
+
+    // Function to move params to top level
+    const moveParamsToTop = (obj: any) => {
+        if (!obj || typeof obj !== 'object') return;
+
+        if (obj.params && Array.isArray(obj.params)) {
+            // Split params into select and non-select
+            const [selectParamsArr, nonSelectParamsArr] = obj.params.reduce(
+                ([select, nonSelect]: [Parameter[], Parameter[]], param: Parameter) => {
+                    if ('select' in param) {
+                        select.push(param);
+                    } else {
+                        nonSelect.push(param);
+                    }
+                    return [select, nonSelect];
+                },
+                [[], []]
+            );
+
+            // Add to respective arrays
+            if (nonSelectParamsArr.length > 0) {
+                params.push(...nonSelectParamsArr);
+            }
+            if (selectParamsArr.length > 0) {
+                selectParams.push(...selectParamsArr);
+            }
+
+            // Remove params from original location
+            delete obj.params;
+        }
+
+        Object.values(obj).forEach(value => {
+            moveParamsToTop(value);
+        });
+    };
+
+    // Move params to their destinations
+    moveParamsToTop(mergedSpec);
+
+    // Add select params to base chart
+    const baseChartLayer = findBaseChartLayer(mergedSpec);
+    if (baseChartLayer && selectParams.length > 0) {
+        baseChartLayer.params = selectParams;
+    }
+    if (params.length > 0) {
+
+        mergedSpec.params = mergeParams(params);
+    }
+
+
+
+   
+    // Apply the extraction to the merged spec
+    
+
+    return  mergedSpec;
+}
+
+function mergeParams(params: any[]): any[] {
+    const paramsByName = new Map<string, any>();
+
+    // Group params by name
+    params.forEach(param => {
+        if (!param.name) return;
+
+        if (paramsByName.has(param.name)) {
+            // Merge with existing param
+            const existing = paramsByName.get(param.name);
+            paramsByName.set(param.name, {
+                ...existing,
+                ...param
+            });
+        } else {
+            paramsByName.set(param.name, param);
+        }
+    });
+
+    return Array.from(paramsByName.values());
+}
