+++ src/components/interactions/index.ts
@@ -1,3 +1,2 @@
-export * from './drag2';
-// export * from './brush'
+export * from './Drag';
 export * from './click';
