+++ src/components/marks/line.ts
@@ -5,7 +5,7 @@ import { compilationContext } from '../../binding/binding';
 import { AnchorProxy, AnchorIdentifer, SchemaType } from "../../types/anchors";
 
 import { generateSignalFromAnchor, createRangeAccessor, generateCompiledValue, generateSignalsFromTransforms, generateSignal } from "../utils";
-import { extractSignalNames } from "../../binding/mergedComponent_CLEAN";
+import { extractSignalNames } from "../../binding/mergedComponent";
 
 const lineBaseContext = {
     "start":{
@@ -98,7 +98,7 @@ const configurations = [{
 
 
 
-import { generateConfigurationAnchors } from "../interactions/drag2";
+import { generateConfigurationAnchors } from "../interactions/Drag";
 
 
 export class Line extends BaseComponent {
