--- src/components/signalFactory.ts
@@ -1,113 +1,0 @@
-
-
-// Signal types
-export interface VegaSignalSpec {
-  name: string;
-  value: any;
-  on: VegaSignalUpdate[];
-}
-
-export interface VegaSignalUpdate {
-  events: any;
-  update: string;
-}
-
-// Constraint types
-export interface Constraint {
-  type: 'expression' | 'absolute' | 'relative';
-  value: string | number;
-  source?: string;
-}
-
-export interface RangeConstraint {
-  start?: Constraint;
-  stop?: Constraint;
-}
-
-// Schema types with clear definitions
-export enum SchemaContainerType {
-  Scalar = 'Scalar',
-  Range = 'Range',
-  Set = 'Set'
-}
-
-export interface ChannelSchema {
-  container: SchemaContainerType;
-  valueType: 'Numeric' | 'Categorical' | 'Boolean';
-  interactive?: boolean;
-}
-
-
-/**
- * Manages creation and updating of Vega signals in a structured way
- */
-class SignalManager {
-    private readonly componentId: string;
-    
-    constructor(componentId: string) {
-      this.componentId = componentId;
-    }
-    
-    /**
-     * Creates a scalar signal configuration
-     */
-    createScalarSignal(
-      channelName: string, 
-      constraints: Constraint[] = [],
-      sourceComponentId?: string
-    ): VegaSignalSpec {
-      const signalName = this.formatSignalName(channelName);
-      
-      // Create base signal spec
-      const signal: VegaSignalSpec = {
-        name: signalName,
-        value: null,
-        on: []
-      };
-      
-      // Add source component update if provided
-      if (sourceComponentId) {
-        signal.on.push({
-          events: { signal: sourceComponentId },
-          update: `${sourceComponentId}.${channelName}`
-        });
-      }
-      
-      // Add constraints as update rules
-      this.addConstraintsToSignal(signal, constraints);
-      
-      return signal;
-    }
-    
-    /**
-     * Creates a range signal (with start and stop values)
-     */
-    createRangeSignal(
-      channelName: string,
-      constraints: RangeConstraint[] = [],
-      sourceComponentId?: string
-    ): VegaSignalSpec[] {
-      const startName = this.formatSignalName(`${channelName}_start`);
-      const stopName = this.formatSignalName(`${channelName}_stop`);
-      
-      // Create both signals with proper references
-      // [implementation details]
-      
-      return [startSignal, stopSignal];
-    }
-    
-    /**
-     * Format a signal name consistently
-     */
-    formatSignalName(channelName: string): string {
-      return `${this.componentId}_${channelName}`;
-    }
-    
-    /**
-     * Add constraints to a signal specification
-     */
-    private addConstraintsToSignal(signal: VegaSignalSpec, constraints: Constraint[]): void {
-      // Process constraints in a clear, structured way
-      // [implementation details]
-    }
-  }
